Move SNP CPUID definitions to a separate crate
