+++ src/cpu/cpuid.rs
@@ -5,45 +5,10 @@
 // Author: Joerg Roedel <jroedel@suse.de>
 
 use crate::utils::immut_after_init::ImmutAfterInitRef;
+use cpuarch::snp_cpuid::SnpCpuidTable;
 use log;
 
-const SNP_CPUID_MAX_COUNT: usize = 64;
-
 static CPUID_PAGE: ImmutAfterInitRef<SnpCpuidTable> = ImmutAfterInitRef::uninit();
-
-#[derive(Copy, Clone, Default, Debug)]
-#[repr(C, packed)]
-pub struct SnpCpuidFn {
-    eax_in: u32,
-    ecx_in: u32,
-    xcr0_in: u64,
-    xss_in: u64,
-    eax_out: u32,
-    ebx_out: u32,
-    ecx_out: u32,
-    edx_out: u32,
-    reserved_1: u64,
-}
-
-#[derive(Copy, Clone, Debug)]
-#[repr(C, packed)]
-pub struct SnpCpuidTable {
-    count: u32,
-    reserved_1: u32,
-    reserved_2: u64,
-    func: [SnpCpuidFn; SNP_CPUID_MAX_COUNT],
-}
-
-impl Default for SnpCpuidTable {
-    fn default() -> Self {
-        SnpCpuidTable {
-            count: Default::default(),
-            reserved_1: Default::default(),
-            reserved_2: Default::default(),
-            func: [SnpCpuidFn::default(); SNP_CPUID_MAX_COUNT],
-        }
-    }
-}
 
 pub fn register_cpuid_table(table: &'static SnpCpuidTable) {
     CPUID_PAGE
